Collect order related fixtures from other files
