+++ saleor/graphql/order/tests/queries/test_draft_order_with_filter.py
@@ -18,49 +18,53 @@ from ....core.connection import where_filter_qs
 from ....tests.utils import get_graphql_content
 from ...filters import OrderDiscountedObjectWhere
 
-
-@pytest.fixture
-def draft_orders_query_with_filter():
-    query = """
-      query ($filter: OrderDraftFilterInput!, ) {
-        draftOrders(first: 5, filter:$filter) {
-          totalCount
-          edges {
-            node {
-              id
-            }
-          }
+DRAFT_ORDERS_QUERY_WITH_FILTER = """
+  query ($filter: OrderDraftFilterInput!, ) {
+    draftOrders(first: 5, filter:$filter) {
+      totalCount
+      edges {
+        node {
+          id
         }
       }
-    """
-    return query
+    }
+  }
+"""
 
 
 def test_draft_orders_query_with_filter_search_by_number(
-    draft_orders_query_with_filter,
     draft_order,
     staff_api_client,
     permission_group_manage_orders,
 ):
+    # given
     update_order_search_vector(draft_order)
     variables = {"filter": {"search": draft_order.number}}
     permission_group_manage_orders.user_set.add(staff_api_client.user)
-    response = staff_api_client.post_graphql(draft_orders_query_with_filter, variables)
+
+    # when
+    response = staff_api_client.post_graphql(DRAFT_ORDERS_QUERY_WITH_FILTER, variables)
     content = get_graphql_content(response)
+
+    # then
     assert content["data"]["draftOrders"]["totalCount"] == 1
 
 
 def test_draft_orders_query_with_filter_search_by_number_with_hash(
-    draft_orders_query_with_filter,
     draft_order,
     staff_api_client,
     permission_group_manage_orders,
 ):
+    # given
     update_order_search_vector(draft_order)
     variables = {"filter": {"search": f"#{draft_order.number}"}}
     permission_group_manage_orders.user_set.add(staff_api_client.user)
-    response = staff_api_client.post_graphql(draft_orders_query_with_filter, variables)
+
+    # when
+    response = staff_api_client.post_graphql(DRAFT_ORDERS_QUERY_WITH_FILTER, variables)
     content = get_graphql_content(response)
+
+    # then
     assert content["data"]["draftOrders"]["totalCount"] == 1
 
 
@@ -76,12 +80,12 @@ def test_draft_order_query_with_filter_customer_fields(
     orders_filter,
     user_field,
     user_value,
-    draft_orders_query_with_filter,
     staff_api_client,
     permission_group_manage_orders,
     customer_user,
     channel_USD,
 ):
+    # given
     setattr(customer_user, user_field, user_value)
     customer_user.save()
     customer_user.refresh_from_db()
@@ -100,11 +104,14 @@ def test_draft_order_query_with_filter_customer_fields(
 
     variables = {"filter": orders_filter}
     permission_group_manage_orders.user_set.add(staff_api_client.user)
-    response = staff_api_client.post_graphql(draft_orders_query_with_filter, variables)
+
+    # when
+    response = staff_api_client.post_graphql(DRAFT_ORDERS_QUERY_WITH_FILTER, variables)
     content = get_graphql_content(response)
+
+    # then
     orders = content["data"]["draftOrders"]["edges"]
     order_id = graphene.Node.to_global_id("Order", order.pk)
-
     assert len(orders) == 1
     assert orders[0]["node"]["id"] == order_id
 
@@ -165,20 +172,23 @@ def test_draft_order_query_with_filter_customer_fields(
 def test_draft_order_query_with_filter_created(
     orders_filter,
     count,
-    draft_orders_query_with_filter,
     staff_api_client,
     permission_group_manage_orders,
     channel_USD,
 ):
+    # given
     Order.objects.create(status=OrderStatus.DRAFT, channel=channel_USD)
     with freeze_time("2012-01-14"):
         Order.objects.create(status=OrderStatus.DRAFT, channel=channel_USD)
     variables = {"filter": orders_filter}
     permission_group_manage_orders.user_set.add(staff_api_client.user)
-    response = staff_api_client.post_graphql(draft_orders_query_with_filter, variables)
-    content = get_graphql_content(response)
-    orders = content["data"]["draftOrders"]["edges"]
 
+    # when
+    response = staff_api_client.post_graphql(DRAFT_ORDERS_QUERY_WITH_FILTER, variables)
+    content = get_graphql_content(response)
+
+    # then
+    orders = content["data"]["draftOrders"]["edges"]
     assert len(orders) == count
 
 
@@ -198,12 +208,12 @@ def test_draft_order_query_with_filter_created(
 def test_draft_orders_query_with_filter_search(
     draft_orders_filter,
     count,
-    draft_orders_query_with_filter,
     staff_api_client,
     permission_group_manage_orders,
     customer_user,
     channel_USD,
 ):
+    # given
     orders = Order.objects.bulk_create(
         [
             Order(
@@ -246,13 +256,16 @@ def test_draft_orders_query_with_filter_search(
         order.search_vector = FlatConcatSearchVector(
             *prepare_order_search_vector_value(order)
         )
+
     Order.objects.bulk_update(orders, ["search_vector"])
-
     variables = {"filter": draft_orders_filter}
-
     permission_group_manage_orders.user_set.add(staff_api_client.user)
-    response = staff_api_client.post_graphql(draft_orders_query_with_filter, variables)
+
+    # when
+    response = staff_api_client.post_graphql(DRAFT_ORDERS_QUERY_WITH_FILTER, variables)
     content = get_graphql_content(response)
+
+    # then
     assert content["data"]["draftOrders"]["totalCount"] == count
