+++ saleor/order/tests/fixtures/__init__.py
@@ -1,2 +1,4 @@
+from .draft_order import *  # noqa: F403
+from .fulfillment import *  # noqa: F403
 from .order import *  # noqa: F403
 from .order_line import *  # noqa: F403
