+++ saleor/order/tests/fixtures/order_line.py
@@ -3,6 +3,8 @@ from decimal import Decimal
 import pytest
 from prices import Money, TaxedMoney
 
+from saleor.order import OrderOrigin
+from saleor.order.fetch import OrderLineInfo
 from saleor.order.models import (
     Order,
 )
@@ -10,9 +12,6 @@ from saleor.warehouse.models import (
     Allocation,
     Stock,
 )
-
-from .... import OrderOrigin
-from ....fetch import OrderLineInfo
 
 
 @pytest.fixture
